refactor: resolve all clippy warnings with proper structural fixes
Eliminate all 138+ clippy warnings across 68 source files, replacing
#[allow(clippy::...)] suppressions with proper structural solutions:

- Replace 9 too_many_arguments suppressions with parameter structs:
  PendingOutcomeInput, ToxicityInputs, OwnFillParams, CrossVenueParams,
  DashboardSnapshotParams, RlAdjustmentParams, CreateOrderParams;
  pass &CheckpointBundle directly; use struct literal + validated()
  for EmissionParams
- Box<Ladder> in Action::Quote to fix large_enum_variant (1216 bytes)
- Fix partial_cmp/cmp ordering in OrderedFloat (was infinite recursion)
- Create TimestampRange type alias and LagAnalyzerStatus struct for
  type_complexity returns
- Merge identical if_same_then_else branches in changepoint detection
- Apply idiomatic Rust patterns: is_multiple_of(), clamp(), enumerate(),
  struct literals with ..Default::default(), derive(Default), to_vec(),
  redundant closure removal, float precision truncation

68 files changed, 974 insertions, 1081 deletions. All 1895 tests pass,
cargo clippy -- -D warnings is clean.

+++ src/market_maker/control/belief.rs
@@ -273,7 +273,7 @@ impl BeliefState {
         let prior = Self::default();
 
         // Fill rate: keep some observations
-        let retained_fills = (self.n_fills as f64 * retention) as f64;
+        let retained_fills = self.n_fills as f64 * retention;
         self.lambda = GammaPosterior::new(
             prior.lambda.alpha + retained_fills,
             prior.lambda.beta + self.total_time * retention,
